+++ src/addon/mod/workshop/pages/submission/submission.ts
@@ -1,4 +1,4 @@
-// (C) Copyright 2015 Moodle Pty Ltd.
+// (C) Copyright 2015 Martin Dougiamas
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -141,7 +141,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Check if we can leave the page or not.
      *
-     * @return Resolved if we can leave it, rejected if not.
+     * @return {boolean|Promise<void>} Resolved if we can leave it, rejected if not.
      */
     ionViewCanLeave(): boolean | Promise<void> {
         const assessmentHasChanged = this.assessmentStrategy && this.assessmentStrategy.hasDataChanged();
@@ -170,7 +170,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Function called when we receive an event of submission changes.
      *
-     * @param data Event data received.
+     * @param {any} data Event data received.
      */
     protected eventReceived(data: any): void {
         if (this.workshopId === data.workshopId) {
@@ -184,7 +184,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Fetch the submission data.
      *
-     * @return Resolved when done.
+     * @return {Promise<void>} Resolved when done.
      */
     protected fetchSubmissionData(): Promise<void> {
         return this.workshopHelper.getSubmissionById(this.workshopId, this.submissionId).then((submissionData) => {
@@ -332,7 +332,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Check if data has changed.
      *
-     * @return True if changed, false otherwise.
+     * @return {boolean} True if changed, false otherwise.
      */
     protected hasEvaluationChanged(): boolean {
         if (!this.loaded || !this.access.canoverridegrades) {
@@ -359,7 +359,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Convenience function to refresh all the data.
      *
-     * @return Resolved when done.
+     * @return {Promise<any>} Resolved when done.
      */
     protected refreshAllData(): Promise<any> {
         const promises = [];
@@ -373,10 +373,6 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
             promises.push(this.workshopProvider.invalidateAssessmentData(this.workshopId, this.assessmentId));
         }
 
-        if (this.assessmentUserId) {
-            promises.push(this.workshopProvider.invalidateReviewerAssesmentsData(this.workshopId, this.assessmentId));
-        }
-
         return Promise.all(promises).finally(() => {
             this.eventsProvider.trigger(AddonModWorkshopProvider.ASSESSMENT_INVALIDATED, this.siteId);
 
@@ -387,7 +383,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Pull to refresh.
      *
-     * @param refresher Refresher.
+     * @param {any} refresher Refresher.
      */
     refreshSubmission(refresher: any): void {
         if (this.loaded) {
@@ -431,7 +427,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Sends the evaluation to be saved on the server.
      *
-     * @return Resolved when done.
+     * @return {Promise<any>} Resolved when done.
      */
     protected sendEvaluation(): Promise<any> {
         const modal = this.domUtils.showModalLoading('core.sending', true);
@@ -465,7 +461,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
      * Perform the submission delete action.
      */
     deleteSubmission(): void {
-        this.domUtils.showDeleteConfirm('addon.mod_workshop.submissiondeleteconfirm').then(() => {
+        this.domUtils.showConfirm(this.translate.instant('addon.mod_workshop.submissiondeleteconfirm')).then(() => {
             const modal = this.domUtils.showModalLoading('core.deleting', true);
             let success = false;
             this.workshopProvider.deleteSubmission(this.workshopId, this.submissionId, this.courseId).then(() => {
@@ -494,7 +490,7 @@ export class AddonModWorkshopSubmissionPage implements OnInit, OnDestroy {
     /**
      * Undo the submission delete action.
      *
-     * @return Resolved when done.
+     * @return {Promise<any>} Resolved when done.
      */
     undoDeleteSubmission(): Promise<any> {
         return this.workshopOffline.deleteSubmissionAction(this.workshopId, this.submissionId, 'delete').finally(() => {
